requirements.txt
# PyTorch 및 관련 라이브러리
torch==2.7.1
torchvision==0.22.1

# 데이터 처리
numpy==1.26.4
pandas>=1.5.0
opencv-python-headless==4.10.0.82
Pillow>=9.0.0
scikit-image>=0.19.0

# 평가 지표
scikit-learn==1.3.2

# ViT 모델
timm==0.4.5

# 설정 관리
pyyaml>=6.0

# 진행 표시
tqdm>=4.64.0

# PEFT (선택사항 - LoRA 사용 시)
# peft>=0.7.0
# transformers>=4.35.0
# accelerate>=0.24.0

# 실험 추적 (선택사항)
# tensorboard>=2.11.0
# wandb>=0.13.0

# 개발 도구 (선택사항)
# ruff>=0.1.0
# pytest>=7.0.0


